Move setAttr to seperate file
